// highscore-client/src/lib/highscore-service/model/fault.ts
/**
 * Hall Of Fame
 * Simple Highscore Application
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Fault { 
    message: string;
    code?: string;
    error?: any;
}